Search is working correctly.
Will need to standardise the layout so that friends always on right and profile always on left and the content in the centre is all that changes.

Speeds are a bit slow for the search, will have to see if there is a better way of searching than just a LIKE comparison.